+++ .storybook/addons.js
@@ -2,5 +2,6 @@
  * @author Luciano Graziani @lgraziani2712
  *
  */
-import '@kadira/storybook/addons';
-import 'storybook-addon-specifications/register';
+import '@storybook/addon-actions/register';
+import '@storybook/addon-links/register';
+// import 'storybook-addon-specifications/register';