+++ .storybook/config.js
@@ -4,12 +4,11 @@
  */
 import 'babel-polyfill';
 
-import { configure } from '@kadira/storybook';
+import { configure } from '@storybook/react';
 
 const req = require.context('../src', true, /\.stories\.jsx$/);
 
 function loadStories() {
-  // require('globalStyles');
   req.keys().forEach(req);
 }