+++ dev-elements/test/storybook-facades.js
@@ -6,23 +6,8 @@
 // $FlowDoNotDisturb internal jest module
 import expectReal from 'jest-matchers';
 
-export {
-  storiesOf,
-  action,
-  linkTo,
-} from '@kadira/storybook';
-
-export {
-  specs,
-  describe,
-  it,
-  beforeEach,
-  before,
-  after,
-  afterEach,
-  xit,
-  fit,
-  xdescribe,
-} from 'storybook-addon-specifications';
+export { storiesOf } from '@storybook/react';
+export { action } from '@storybook/addon-actions';
+export { linkTo } from '@storybook/addon-links';
 
 export const expect = expectReal;