+++ src/blockly/components/ActionContainer.js
@@ -5,7 +5,7 @@
  * @flow
  */
 import * as colors from 'constants/colors';
-import { ACTION_CONTAINER } from 'constants/actions';
+import { ACTION_CONTAINER } from 'constants/instructions';
 
 const mainInput = 'program';
 
@@ -14,6 +14,7 @@ Blockly.Blocks[ACTION_CONTAINER] = {
     this.appendDummyInput()
         .appendField(Blockly.Msg.ACTION_CONTAINER_TITLE);
     this.appendStatementInput(mainInput);
+    this.contextMenu = false;
     this.setDeletable(false);
     this.setEditable(false);
     this.setMovable(false);