+++ src/blockly/components/LeaveMaze.js
@@ -5,16 +5,17 @@
  * @flow
  */
 import * as colors from 'constants/colors';
-import { LEAVE_MAZE } from 'constants/actions';
+import { LEAVE_MAZE } from 'constants/instructions';
 
 Blockly.Blocks[LEAVE_MAZE] = {
   init: function init() {
     this.appendDummyInput()
         .setAlign(Blockly.ALIGN_CENTRE)
         .appendField(Blockly.Msg.LEAVE_MAZE_TITLE);
-    this.setPreviousStatement(true, null);
-    this.setNextStatement(true, null);
+    this.contextMenu = false;
     this.setColour(colors.ACTION);
+    this.setNextStatement(true, null);
+    this.setPreviousStatement(true, null);
     this.setTooltip(Blockly.Msg.LEAVE_MAZE_TOOLTIP);
   },
 };