+++ src/blockly/components/MoveRight.js
@@ -5,16 +5,17 @@
  * @flow
  */
 import * as colors from 'constants/colors';
-import { MOVE_RIGHT } from 'constants/actions';
+import { MOVE_RIGHT } from 'constants/instructions';
 
 Blockly.Blocks[MOVE_RIGHT] = {
   init: function init() {
     this.appendDummyInput()
         .setAlign(Blockly.ALIGN_CENTRE)
         .appendField(Blockly.Msg.MOVE_RIGHT_TITLE);
-    this.setPreviousStatement(true, null);
-    this.setNextStatement(true, null);
+    this.contextMenu = false;
     this.setColour(colors.ACTION);
+    this.setNextStatement(true, null);
+    this.setPreviousStatement(true, null);
     this.setTooltip(Blockly.Msg.MOVE_RIGHT_TOOLTIP);
   },
 };